improve console error page (close #272) (#1890)
