--- console/src/components/App/ErrorBoundary.js
@@ -1,30 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-class ErrorBoundary extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { hasError: false, info: null };
-  }
-
-  componentDidCatch(error, info) {
-    // Display fallback UI
-    this.setState({ hasError: true, info: info });
-    // You can also log the error to an error reporting service
-    // logErrorToMyService(error, info);
-  }
-
-  render() {
-    if (this.state.hasError) {
-      // You can render any custom fallback UI
-      return <div>Something went wrong</div>;
-    }
-    return this.props.children;
-  }
-}
-
-ErrorBoundary.propTypes = {
-  children: PropTypes.element,
-};
-
-export default ErrorBoundary;
