+++ console/src/components/App/ErrorPage.scss
@@ -2,7 +2,7 @@
   padding: 0;
 }
 
-.viewcontainer {
+.viewContainer {
   height: 100vh;
   width: 100vw;
   display: table;
