+++ console/src/components/index.js
@@ -4,5 +4,5 @@
 
 export App from './App/App';
 export Main from './Main/Main';
-export PageNotFound from './404/PageNotFound';
+export PageNotFound from './Error/PageNotFound';
 export Login from './Login/Login';
